Found the source of the icon bug. I was attempting to load a float result into a uint32 field. I really need to fix intellisense, because a silly thing like that would have most likely been caught by it.
Former-commit-id: 68c99faf394f9cc01f6c3ecc82265c2f1d546738

+++ Source/Crystalline/Public/CrystallineHUD.h
@@ -21,10 +21,10 @@ public:
 
 private:
 	/** The vertical scale factor of the UI Relative to 1080.*/
-	uint32 ScaleUIY;
+	float ScaleUIY;
 
 	/** The horizontal scale factor of the UI Relative to 1920.*/
-	//uint32 ScaleUIX; Commented for potential future usage. -John
+	//float ScaleUIX; Commented for potential future usage. -John
 
 };
